requirements.txt
appdirs==1.4.2
Django==1.10.6
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
six==1.10.0
django-import-export
